Remove metrics POST request from Interceptor.py, add application label to PrometheusMetrics in app.py, Jenkinsfile updated
The Interceptor.py code was changed to remove POST requests to the metrics-service, as these are now made redundant by the use of PrometheusMetrics. The addition of an application label to PrometheusMetrics in app.py allows for easier monitoring and tracking of application metrics. The Jenkinsfile has been updated to use sshCommand instead of a simple shell command for security reasons. It also now uses credentials stored in the environment for authenticating the connection to the remote server, in line with best security practices. This adds a further layer of protection against potential security breaches.