Add ScriptsService and modify Utils.py, app.py, requirements.txt
A new functionality was added to execute SQL scripts through an HTTP endpoint. ScriptsService.py was introduced to handle request properties and execute scripts. Utils.py was modified to read scripts from file. The related endpoint was added to app.py. New dependencies were added to requirements.txt.

The change was necessary to provide an interface to execute SQL scripts remotely in a controlled manner, which could be useful for ad-hoc data analysis and debug procedures. This function assumes scripts follow a certain parameter pattern and default dates are provided if no range is passed.