Add 'with_uncategorized' argument to get_dataset function
The 'get_dataset' function in DatasetService.py now accepts a new 'with_uncategorized' argument. This flag determines whether to group uncategorized data with the rest of the data or not. The default value set is False, providing more flexibility to the user. There are also minor changes in import statements.